Move the former human thingcomps to the thingcomps namespace
